Move update logic to entity
